fixed masking

+++ run_env_ppo.py
@@ -69,7 +69,7 @@ def main():
         # logging
         if i_episode % cf.log_interval == 0:
             avg_length = int(avg_length / cf.log_interval)
-            running_reward = int((running_reward / cf.log_interval))
+            running_reward = (running_reward / cf.log_interval)
 
             print('Episode {} \t Avg length: {} \t Avg reward: {}'.format(i_episode, avg_length, running_reward))
             running_reward = 0
